Framework: Rename boot/index.js to use project specific name
